A tweek to the Check Required Fields JavaScript code.
Clean up the View Audit Trail code, so it is easier to read, and "createObject"
is called (as God intended).

Change the Set Password code so auditing actually works (the userInfo was not
being recorded before) and the logged in user is accessed using sane and
sensible means.

Fix up the event-ID generator so there is more food for the MD5 hash.

--HG--
extra : convert_revision : svn%3A01e103ed-f575-20b1-d296-6440ccab0075/branches/GSAuditTrail/trunk%404478